Add ModdedPA version to device info settings
We should have the same changes for a new build, to avoid confuses.

I'll add a line "ro.moddedpa.version=<version>" into the vendor/pa/version.mk file, so when we build with it the code should add in automatically to build.prop file. And it will show up in the device info settings.