Fix navring targets build
The resources are copied together from the Keyguard Package and the cyanogenmod repo. Thanks for their hard work!

Navring targets works now, the only problem I have is, that the animation of the quicklaunch ring is a bit laggy but I don't know whether this is caused by the navring commits or anything else.
And another thing: I didn't added the "Utils.isPhone" function from cyanogenmod which is needed to detect whether a large screen is used or not. So if the settings lock weird on a tablet it's probably caused by that and we have to fix it.